requirements.txt
transformers
langchain
streamlit
python-dotenv
openai



